Change locations to be queryable by their constituent levels
This replaces the current behaviour (we also do this in EES) of treating each location as an
aggregate of multiple locations at different levels (otherwise known as 'attributes').

We've changed each 'attribute' to be its own location, meaning that users should now be able to
query locations in a more intuitive way.